requirements.txt
pyqt5==5.15.10
pyserial==3.5
matplotlib==3.9.0


